Refactor Search to be `client:only`
Render the button using Astro and hook up the component on the client
