+++ src/components/Header/DocSearch.css
@@ -1,0 +1,23 @@
+:root {
+	--docsearch-hit-active-color: var(--theme-text);
+	--docsearch-primary-color: var(--theme-text-accent);
+	--docsearch-logo-color: var(--theme-text);
+}
+
+.DocSearch-Modal .DocSearch-Hit a {
+	box-shadow: none;
+	border: 1px solid var(--theme-glow-diffuse);
+}
+
+/* This query matches the upstream media query for mobile devices */
+@media (max-width: 750px) {
+	/* Increase spacing to match our site layout */
+	:root {
+		--docsearch-spacing: 15px;
+	}
+
+	/* Fix search not being displayed correctly after scrolling down */
+	.DocSearch-Container {
+		position: fixed;
+	}
+}
