+++ src/components/Header/Header.astro
@@ -2,7 +2,7 @@
 import SkipToContent from './SkipToContent.astro';
 import SidebarToggle from './SidebarToggle';
 import LanguageSelect from './LanguageSelect';
-import Search from './Search';
+import Search from './Search.astro';
 import ThemeToggleButton from './ThemeToggleButton';
 import { getLanguageFromURL } from '../../util';
 import { getDocSearchStrings, useTranslations } from '../../i18n/util';
