+++ src/components/Header/Search.astro
@@ -1,0 +1,77 @@
+---
+import '@docsearch/css/dist/style.css';
+import { DocSearchTranslation } from "../../i18n/translation-checkers";
+import DocSearch from './DocSearch';
+import './HeaderButton.css';
+
+export interface Props {
+	lang?: string;
+	labels: DocSearchTranslation;
+}
+
+const { lang, labels } = Astro.props as Props;
+const { button, shortcutLabel, modal, placeholder } = labels;
+---
+
+<button class="header-button search-input" type="button" aria-label={button} id="docsearch-search-button">
+	<svg width="1em" height="1em" viewBox="0 0 24 24" fill="none" focusable="false" aria-hidden="true" role="img">
+		<path d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z" stroke="currentColor" stroke-width="2" stroke-linecap="round"
+			stroke-linejoin="round" transform="translate(-1)" />
+	</svg>
+	<span class="search-placeholder">
+		{button} <span class="sr-only">({shortcutLabel})</span>
+	</span>
+	<span class="search-keyboard-hint">
+		<kbd aria-hidden="true">/</kbd>
+	</span>
+</button>
+<DocSearch {...{ lang, labels: { modal, placeholder } }} client:only="preact" />
+
+<style>
+	/** Style search header button */
+	.search-input {
+		position: relative;
+		display: flex;
+	}
+
+	.search-placeholder {
+		text-align: initial;
+		margin-inline-start: 0.25rem;
+		margin-inline-end: 3rem;
+		display: none;
+	}
+
+	.search-keyboard-hint {
+		position: absolute;
+		inset-inline-end: 0.5rem;
+		padding: 3px 5px;
+		display: none;
+		align-items: center;
+		justify-content: center;
+		font-size: 13px;
+		font-family: var(--font-mono);
+		letter-spacing: 0.125rem;
+		line-height: 14px;
+		pointer-events: none;
+		border-color: var(--theme-shade-subtle);
+		border-style: solid;
+		border-width: 1px;
+		border-radius: 0.25rem;
+	}
+
+	@media (min-width: 50em) {
+		.search-input {
+			--background-color-default: var(--theme-divider);
+			--border-color-default: var(--theme-divider);
+			min-width: 10rem;
+		}
+
+		.search-placeholder {
+			display: flex;
+		}
+
+		.search-keyboard-hint {
+			display: flex;
+		}
+	}
+</style>
