--- src/components/Header/Search.css
@@ -1,67 +1,0 @@
-/** Style Algolia */
-:root {
-	--docsearch-hit-active-color: var(--theme-text);
-	--docsearch-primary-color: var(--theme-text-accent);
-	--docsearch-logo-color: var(--theme-text);
-}
-
-.DocSearch-Modal .DocSearch-Hit a {
-	box-shadow: none;
-	border: 1px solid var(--theme-glow-diffuse);
-}
-
-/* This query matches the upstream media query for mobile devices */
-@media (max-width: 750px) {
-	/* Increase spacing to match our site layout */
-	:root {
-		--docsearch-spacing: 15px;
-	}
-
-	/* Fix search not being displayed correctly after scrolling down */
-	.DocSearch-Container {
-		position: fixed;
-	}
-}
-
-/** Style search header button */
-.search-input {
-	position: relative;
-	display: flex;
-}
-.search-placeholder {
-	text-align: initial;
-	margin-inline-start: 0.25rem;
-	margin-inline-end: 3rem;
-	display: none;
-}
-.search-keyboard-hint {
-	position: absolute;
-	inset-inline-end: 0.5rem;
-	padding: 3px 5px;
-	display: none;
-	align-items: center;
-	justify-content: center;
-	font-size: 13px;
-	font-family: var(--font-mono);
-	letter-spacing: 0.125rem;
-	line-height: 14px;
-	pointer-events: none;
-	border-color: var(--theme-shade-subtle);
-	border-style: solid;
-	border-width: 1px;
-	border-radius: 0.25rem;
-}
-
-@media (min-width: 50em) {
-	.search-input {
-		--background-color-default: var(--theme-divider);
-		--border-color-default: var(--theme-divider);
-		min-width: 10rem;
-	}
-	.search-placeholder {
-		display: flex;
-	}
-	.search-keyboard-hint {
-		display: flex;
-	}
-}
